# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Anby Wiki API
  version: "0.1.0"
  description: |
    Anby Wiki HTTP API 权威契约（OpenAPI 3.1）。
    本文件是所有接口行为的唯一权威定义：错误模型、分页、幂等头与请求 ID 在此统一定义。
    前端客户端由 OpenAPI Generator 从本文件生成，禁止手写 URL/DTO。
    Go 侧 handler 手写实现，由 Contract Test 保证与本文件一致。
  contact:
    name: Anby Wiki
...
servers:
  - url: http://localhost:3000
    description: 本地 Web 入口（Next.js 将 /api/* 转发到 Go API）
  - url: http://localhost:8080
    description: 本地直连 Go API
...
paths:
...
  /api/v1/pages/{id}:
    get:
      operationId: getPageByID
      summary: 按 ID 读取页面当前版本
      description: |
        阅读端点（匿名可读，无需登录）。按页面 ID 读取当前 Revision 与渲染 HTML，
        响应结构与行为同 getPageByTitle（重定向同样跟随；软删除页返回 410 gone；
        未发布过 content 为 null）。
      tags: [reading]
      parameters:
        - $ref: "#/components/parameters/PageId"
        - $ref: "#/components/parameters/ContentMode"
      responses:
        "200":
          description: 页面元信息 + 当前 Revision 元信息 + AST + 渲染 HTML
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageWithContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/pages/{id}/backlinks:
    get:
      operationId: listBacklinks
      summary: 页面反向链接
      description: |
        投影查询端点（匿名可读，无需登录）。列出指向该页面的已解析页面引用来源：
        来源页（id/标题）+ 所在 Block + 展示文本。数据来自 page_link_projection 投影表
        （设计 §17.3：关系查询走投影表，不扫 AST），由 Worker 异步构建，
        新发布内容与投影之间存在最终一致窗口。
        按 (source_page_id, source_block_id, source_node_id) 升序游标分页：
        首页不传 cursor；后续页传上一页响应的 next_cursor；next_cursor 为 null 表示没有更多。
        游标无法解析返回 400 validation_failed；页面不存在返回 404；软删除页返回 410 gone。
      tags: [projection]
      parameters:
        - $ref: "#/components/parameters/PageId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 一页反向链接
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacklinkListPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/entities/{id}/mentions:
    get:
      operationId: listEntityMentions
      summary: Entity 页面提及位置
      description: |
        匿名反向查询 EntityReference 的页面使用位置。只读 entity_mention_projection，
        不扫描 AST JSON；每条结果携带来源 Revision、Block、Node 与 mention_text。
      tags: [projection]
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 一页 Entity 提及位置
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceUsageListPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/claims/{id}/usages:
    get:
      operationId: listClaimUsages
      summary: Claim 页面使用位置
      description: |
        匿名反向查询 ClaimReference 的页面使用位置。只读 claim_usage，不扫描 AST JSON；
        每条结果携带来源 Revision、Block 与 Node。
      tags: [projection]
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 一页 Claim 使用位置
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceUsageListPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/citations/{id}/usages:
    get:
      operationId: listCitationUsages
      summary: Citation 页面使用位置
      description: |
        匿名反向查询 CitationReference 的页面使用位置。只读 citation_usage，不扫描 AST JSON；
        每条结果携带来源 Revision、Block、Node 与可选 Claim 上下文。
      tags: [projection]
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 一页 Citation 使用位置
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceUsageListPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/sources/{id}/usages:
    get:
      operationId: listSourceUsages
      summary: Source 页面使用位置
      description: |
        匿名反向查询一个 Source 经 SourceVersion、Citation 被当前页面使用的位置。
        只读 citation_usage 及证据身份链，不扫描 AST JSON。
      tags: [projection]
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 一页 Source 使用位置
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceUsageListPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/components/{id}/usages:
    get:
      operationId: listComponentUsages
      summary: Component 页面依赖位置
      description: |
        匿名反向查询当前页面引用一个 Component 的位置。只读
        component_dependency，不扫描 AST JSON；可按不可变版本过滤。
      tags: [projection]
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
        - name: version
          in: query
          schema:
            type: integer
            minimum: 1
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 一页 Component 使用位置
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentUsageListPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: anby_session
      description: HttpOnly 同源服务端 session cookie；名称可由部署配置覆盖
  headers:
    X-Request-ID:
      description: 请求 ID；客户端可携带同名请求头，服务端透传或重新生成
      schema:
        type: string
  parameters:
    PageId:
      name: id
      in: path
      required: true
      description: 页面 ID（UUIDv7）
      schema:
        type: string
        format: uuid
    ReferenceTargetId:
      name: id
      in: path
      required: true
      description: Entity、Claim 或 Citation 稳定 ID
      schema:
        type: string
        format: uuid
...
    ContentMode:
      name: content_mode
      in: query
      required: false
      description: >-
        auto 对大于 64 KiB 且分片投影就绪的页面返回 sections；full 强制全文；
        sections 在投影就绪时强制分片，否则安全降级为 full。
      schema:
        type: string
        enum: [auto, full, sections]
        default: auto
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: 上一页响应返回的 next_cursor；首页不传
      schema:
        type: string
    PageSize:
      name: page_size
      in: query
      required: false
      description: 每页条数，默认 20，最大 100
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Page:
      type: object
      required:
        - id
        - wiki_id
        - namespace_id
        - normalized_title
        - display_title
        - language
        - content_model
        - status
        - current_revision_id
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        wiki_id:
          type: string
          format: uuid
        namespace_id:
          type: string
          format: uuid
        normalized_title:
          type: string
          description: 规范化标题（路由唯一性键之一）
          example: anby demara
        display_title:
          type: string
          description: 展示标题
          example: Anby Demara
        language:
          type: string
          example: zh-Hans
        content_model:
          type: string
          example: block-v1
        status:
          type: string
          example: active
        current_revision_id:
          type: string
          format: uuid
          nullable: true
          description: 当前 Revision；未发布过为 null
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PageRedirectTarget:
      type: object
      additionalProperties: false
      required: [kind]
      description: >-
        判别式重定向目标。page 需要 target_page_id；page_section 还需要
        anchor_block_id；unresolved 需要 namespace 与 target_title；
        interwiki 需要 target_title 与 external_url。
      properties:
        kind:
          type: string
          enum: [page, page_section, unresolved, interwiki]
        target_page_id:
          type: string
          format: uuid
        target_page_title:
          type: string
          readOnly: true
        namespace_id:
          type: string
          format: uuid
          readOnly: true
        namespace:
          type: string
          minLength: 1
          maxLength: 64
        target_title:
          type: string
          minLength: 1
          maxLength: 255
        anchor_block_id:
          type: string
          format: uuid
        external_url:
          type: string
          format: uri
          maxLength: 2048
...
    PageWithContent:
      type: object
      required: [page, content, via_alias]
      description: 阅读端点响应：落地页元信息 + 当前 Revision 内容（渲染直连 ContentSnapshot）
      properties:
        page:
          $ref: "#/components/schemas/Page"
        content:
          nullable: true
          description: 当前 Revision 内容；页面已创建但未发布过为 null
          $ref: "#/components/schemas/PageContent"
        via_alias:
          type: boolean
          description: 是否经由 page_alias（旧标题别名）解析命中
        alias_title:
          type: string
          description: via_alias=true 时回显客户端请求的标题原文
        redirect:
          type: object
          nullable: true
          description: >-
            请求命中重定向源时出现。resolved=true 时 page/content 是本地落地页；
            unresolved 或 interwiki 终点返回 resolved=false、content=null，并由 target
            给出可治理的目标。
          required: [from_page_id, from_title, target, resolved, hops]
          properties:
            from_page_id:
              type: string
              format: uuid
              description: 重定向源页面 ID
            from_title:
              type: string
              description: 重定向源页面显示标题
            target:
              $ref: "#/components/schemas/PageRedirectTarget"
            resolved:
              type: boolean
            hops:
              type: integer
              minimum: 1
    PageContent:
      type: object
      required:
        [revision, delivery_mode, section_count, ast_json, html, renderer_version]
      description: >-
        已发布页面的当前 Revision 内容。delivery_mode=sections 时全文 AST/HTML
        为 null，客户端按 PageSectionManifest 加载可重建分片。
      properties:
        revision:
          $ref: "#/components/schemas/Revision"
        delivery_mode:
          type: string
          enum: [full, sections]
        section_count:
          type: integer
          minimum: 1
        ast_json:
          type: object
          nullable: true
          additionalProperties: true
          description: |
            当前 Revision 的 canonical Typed Block AST
            （权威 Schema：contracts/schemas/ast/v1/ast.schema.json）。
        html:
          type: string
          nullable: true
          description: 由 ast_json 渲染的安全 HTML 片段（不含 html/head；安全规则见 backend/internal/render/README.md）
        renderer_version:
          type: string
          description: 渲染器版本（M3 RenderedPage 投影按此版本判断缓存重建）
          example: v1
    Backlink:
      type: object
      required: [source_page_id, source_title, source_block_id, display_text]
      description: 一条反向链接：指向目标页的已解析引用来源
      properties:
        source_page_id:
          type: string
          format: uuid
          description: 来源页面 ID
        source_title:
          type: string
          description: 来源页面显示标题
          example: Anby Demara
        source_block_id:
          type: string
          format: uuid
          description: 引用所在 Block ID
        display_text:
          type: string
          description: 引用的展示文本
    BacklinkListPage:
      allOf:
        - $ref: "#/components/schemas/CursorPage"
        - type: object
          required: [items, next_cursor]
          description: 反向链接列表（source_page_id, source_block_id, source_node_id 升序游标分页）
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Backlink"
            next_cursor:
              type: string
              nullable: true
              description: 下一页游标；null 表示没有更多
...
    ReferenceUsage:
      type: object
      required: [page_id, page_title, revision_id, block_id, node_id, mention_text, claim_id]
      description: Entity/Claim/Citation 在页面 Current Revision 中的一处引用位置
      properties:
        page_id:
          type: string
          format: uuid
        page_title:
          type: string
        revision_id:
          type: string
          format: uuid
          description: 产生该使用投影的来源 Revision
        block_id:
          type: string
          format: uuid
        node_id:
          type: string
          description: 引用节点在 Block content 内的稳定定位键
        mention_text:
          type: string
          nullable: true
          description: EntityReference 的展示文本；Claim/Citation 使用位置为 null
        claim_id:
          type: string
          format: uuid
          nullable: true
          description: Citation 的可选 Claim 上下文；直接 CitationReference 为 null
    ReferenceUsageListPage:
      allOf:
        - $ref: "#/components/schemas/CursorPage"
        - type: object
          required: [items, next_cursor]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ReferenceUsage"
            next_cursor:
              type: string
              nullable: true
              description: 下一页游标；null 表示没有更多
    SourceUsage:
      type: object
      required:
        - page_id
        - page_title
        - revision_id
        - block_id
        - node_id
        - citation_id
      properties:
        page_id: { type: string, format: uuid }
        page_title: { type: string }
        revision_id:
          type: string
          format: uuid
          description: 产生该投影的当前 Revision
        block_id: { type: string, format: uuid }
        node_id: { type: string }
        citation_id: { type: string, format: uuid }
        claim_id:
          type: string
          format: uuid
          nullable: true
          description: 通过 Claim 展示证据时的上下文；正文直接引用时为空
    SourceUsageListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SourceUsage" }
        next_cursor:
          type: string
          nullable: true
    ComponentUsage:
      type: object
      required:
        - page_id
        - page_title
        - revision_id
        - block_id
        - component_version
        - entity_id
      properties:
        page_id: { type: string, format: uuid }
        page_title: { type: string }
        revision_id:
          type: string
          format: uuid
          description: 产生该投影的当前 Revision
        block_id: { type: string, format: uuid }
        component_version: { type: integer, minimum: 1 }
        entity_id: { type: string, format: uuid }
    ComponentUsageListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ComponentUsage" }
        next_cursor:
          type: string
          nullable: true
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: 稳定机器可读错误码
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - gone
            - conflict
            - stale_revision
            - validation_failed
            - rate_limited
            - internal
        message:
          type: string
          description: 面向调用者的错误描述（不泄露内部实现细节）
        request_id:
          type: string
          description: 与响应头 X-Request-ID 一致，用于日志串联
        details:
          type: object
          description: 可选结构化错误明细（如字段级校验错误）
          additionalProperties: true
    CursorPage:
      type: object
      required: [items, next_cursor]
      description: 统一游标分页信封；列表端点以内联 allOf 方式复用
      properties:
        items:
          type: array
          items: {}
        next_cursor:
          type: string
          nullable: true
          description: 下一页游标；null 表示没有更多
...
  responses:
    BadRequest:
      description: 请求格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Gone:
      description: 资源曾存在但已删除（如软删除页面、重定向目标已删除）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: 请求语义可理解但无法处理（如重定向环/重定向链过深）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Internal:
      description: 服务端内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"